init-cluster saves source info to target
Failures in steps after prepare init-cluster were failing.  Root cause
is that we were not switching the dbConnection to the newly minted
cluster before we query and and save off the target cluster's
configuration information.

Our next move was to pull the async bit into a helper that can be tested
in isolation.  We were stopped struggling against the desire not to add
something else to the hub at this commit, because InitCluster() is
deeply is to intwined with hub.

The choice to have cluster.NewDBconn not connect, implies that callers
will need to connect (and then defer a close).  Most of the fallout from
that decision shows up in tests that assumed the database connection had
connected, and figured out which version of the DB that connection had
been made to.

Co-authored-by: Jim Doty <jdoty@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>